Added class based views

+++ restaurantList_app/api/views1.py
@@ -1,43 +1,40 @@
-from rest_framework.response import Response
-from rest_framework import status
-from rest_framework.views import APIView
-
-
-from restaurantList_app.models import Restaurant, Location
+from restaurantList_app.models import Restaurant
 from restaurantList_app.api.serializers import RestaurantSerializer
+from rest_framework.response import Response
+from rest_framework.decorators import api_view
+from rest_framework import status
 
-
-
-class RestarauntListAV(APIView):
+@api_view(http_method_names=['GET', 'POST'])
+def restra_list(request):
     
-    def get(self, request):
+    if request.method == 'GET' :
         restras = Restaurant.objects.all()
         serializer = RestaurantSerializer(restras, many=True)
         return Response(serializer.data)
-        
-    def post(self, request):
-        
-        serializer =RestaurantSerializer(data = request.data)
-        if serializer.is_valid():
+    
+    if request.method == 'POST' :
+        serializer = RestaurantSerializer(data = request.data)
+        if serializer.is_valid() :
             serializer.save()
             return Response(serializer.data)
-        
         else:
             return Response(serializer.errors)
-        
-        
-class RestaurantDetailAV(APIView):
-    def get(self, request, pk):
-        try:
+    
+    
+
+@api_view(http_method_names=['GET', 'PUT', 'DELETE'])
+def restra_details(request, pk):
+    
+    if request.method == 'GET':
+        try :
             restra = Restaurant.objects.get(pk=pk)
         except Restaurant.DoesNotExist:
-            return Response({'error' :'Not found'}, status=status.HTTP_404_NOT_FOUND)
+            return Response({'error' :'Restaurant Not Found'},status=status.HTTP_404_NOT_FOUND)
         
         serializer = RestaurantSerializer(restra)
         return Response(serializer.data)
-        
-        
-    def put(self, request, pk):
+    
+    if request.method == 'PUT':
         restra = Restaurant.objects.get(pk=pk)
         serializer = RestaurantSerializer(restra, data=request.data)
         if serializer.is_valid():
@@ -45,11 +42,9 @@ class RestaurantDetailAV(APIView):
             return Response(serializer.data)
         else :
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-    def delete(self, request, pk):
+    
+    if request.method == 'DELETE':
         restra = Restaurant.objects.get(pk=pk)
         restra.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
-
-
